MediTrack/BackEnd/requirements.txt
fastapi
uvicorn
python-multipart
supabase
python-dotenv
pandas
pydantic
pydantic-settings
cryptography
qrcode
Pillow
redis
google-generativeai
httpx
python-jose[cryptography]
passlib[bcrypt]
bcrypt
aiofiles
reportlab
pydantic[email]
aiohttp
pytesseract
Pillow
opencv-python-headless
